Move copy files to shapefile level.
